Change the process of selecting elements relative to another
Previously selection methods would automatically select relative to
a selected element as long as one was selected and not null. This
meant before traversing the whole DOM selected elements would have to be
cleared. Forgetting to do this can easily cause confusion when elements
were not found.

The new method of relative selection is to do so explicitly by adding
the optional matching string to the end of selection step definitions
'relative to the selected element'. If this is not included then all
selection methods will traverse the DOM from the root by default.

- Add examples for relative selection
- Update readme and documentation to reflect changes